backend/requirements.txt
# API & async
fastapi==0.109.2
uvicorn[standard]==0.27.1
python-multipart==0.0.9

# DB
sqlalchemy[asyncio]==2.0.25
asyncpg==0.29.0
psycopg2-binary==2.9.9
alembic==1.13.1

# Redis & Celery
redis==5.0.1
celery[redis]==5.3.6

# Auth
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4

# Storage (S3-compatible)
boto3==1.34.34

# Document processing (use Python 3.11 for pre-built PyMuPDF wheels on Windows)
pymupdf==1.24.1
pytesseract==0.3.10
Pillow==10.2.0

# Excel/Word/PPT generation
openpyxl==3.1.2
python-docx==1.1.0
python-pptx==0.6.23

# LLM (optional)
openai==1.12.0
httpx==0.26.0

# Utils
pydantic==2.6.1
pydantic-settings==2.1.0
python-dotenv==1.0.1
structlog==24.1.0

# Observability
sentry-sdk[fastapi]==1.40.0
opentelemetry-api==1.22.0
opentelemetry-sdk==1.22.0
